# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MyReadings API
  description: >
    API REST de MyReadings: gestión de lecturas familiares con sistema de
    recompensas. Roles PARENT (padre/madre) y CHILD (hijo/a). Todos los
    endpoints salvo `/auth/register` y `/auth/login` requieren JWT
    (`Authorization: Bearer <token>`).
  version: "1.0.0"
  contact:
    name: MyReadings (TFM)
servers:
  - url: http://localhost:3000/api
    description: Desarrollo local
  - url: /api
    description: Relativo (mismo host que el frontend)
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Registro de un padre/madre
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Padre/madre registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      tags: [auth]
      summary: Login (padre o hijo)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login correcto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [auth]
      summary: Perfil del usuario autenticado
      responses:
        "200":
          description: Usuario autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeUser"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─────────────────────────── USERS (children) ───────────────────────────
...
  /categories:
    get:
      tags: [categories]
      summary: Listar categorías visibles
      description: >
        Devuelve las categorías propias. Si el usuario es CHILD, incluye
        además (en modo solo-lectura) las categorías de su padre.
      responses:
        "200":
          description: Lista de categorías
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [categories]
      summary: Crear una categoría propia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Categoría creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /categories/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [categories]
      summary: Categoría por id
      description: Propia, o del padre si el actor es CHILD.
      responses:
        "200":
          description: Categoría
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [categories]
      summary: Actualizar una categoría (solo propietario)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Categoría actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [categories]
      summary: Eliminar una categoría (solo propietario)
      description: 409 si la categoría todavía tiene libros asociados.
      responses:
        "204":
          description: Eliminada
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ─────────────────────────── BOOKS ───────────────────────────
  /books:
    get:
      tags: [books]
      summary: Listar libros propios
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/BookStatus"
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de libros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [books]
      summary: Crear un libro propio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"
      responses:
        "201":
          description: Libro creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /books/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [books]
      summary: Detalle de un libro propio
      responses:
        "200":
          description: Libro
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [books]
      summary: Editar un libro / cambiar de estado
      description: "Validación: `endDate >= startDate` cuando ambas existen."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookRequest"
      responses:
        "200":
          description: Libro actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [books]
      summary: Eliminar un libro propio
      responses:
        "204":
          description: Eliminado
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /children/{childId}/goals:
    parameters:
      - name: childId
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [goals]
      summary: Listar metas de un hijo
      description: Solo PARENT; el hijo debe ser suyo.
      responses:
        "200":
          description: Metas del hijo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [goals]
      summary: Crear una meta para un hijo
      description: Solo PARENT; el hijo debe ser suyo; `targetPoints > 0`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Meta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /goals/mine:
    get:
      tags: [goals]
      summary: Mis metas y progreso (hijo)
      responses:
        "200":
          description: Lista de metas propias con progreso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalWithProgress"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /goals/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    patch:
      tags: [goals]
      summary: Editar una meta
      description: Solo PARENT; el hijo debe ser suyo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalRequest"
      responses:
        "200":
          description: Meta actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [goals]
      summary: Eliminar una meta
      description: Solo PARENT; hijo propio; sin puntos comprometidos.
      responses:
        "204":
          description: Eliminada
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      example: 1
  responses:
    BadRequest:
      description: Validación de datos fallida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Rol o propiedad insuficiente
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflicto (email duplicado, meta ya canjeada, categoría con libros, etc.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: "El color debe ser un hex #RRGGBB"
        error:
          type: string
          example: "Bad Request"
    Role:
      type: string
      enum: [PARENT, CHILD]
    BookStatus:
      type: string
      enum: [NOT_STARTED, READING, FINISHED]
    GoalStatus:
      type: string
      enum: [ACTIVE, ACHIEVED, REDEEMED]
...
    SafeUser:
      type: object
      description: Proyección de usuario sin `passwordHash`.
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "María García"
        email:
          type: string
          format: email
          example: "maria@familia.dev"
        role:
          $ref: "#/components/schemas/Role"
        parentId:
          type: integer
          nullable: true
          example: null
      required: [id, name, email, role, parentId]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          example: "María García"
        email:
          type: string
          format: email
          example: "maria@familia.dev"
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          example: "password123"
      required: [name, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "maria@familia.dev"
        password:
          type: string
          format: password
          example: "password123"
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: "#/components/schemas/SafeUser"
      required: [accessToken, user]
...
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 5
        ownerUserId:
          type: integer
          example: 1
        title:
          type: string
          example: "Ciencia ficción"
        description:
          type: string
          nullable: true
          example: "Novela SF y fantasía"
        colorHex:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          example: "#1E88E5"
        createdAt:
          type: string
          format: date-time
      required: [id, ownerUserId, title, colorHex, createdAt]
    CreateCategoryRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          example: "Ciencia ficción"
        colorHex:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          example: "#1E88E5"
        description:
          type: string
          maxLength: 500
          example: "Novela SF y fantasía"
      required: [title, colorHex]
    UpdateCategoryRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        colorHex:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        description:
          type: string
          maxLength: 500

    # ---------- Books ----------
    Book:
      type: object
      properties:
        id:
          type: integer
          example: 10
        ownerUserId:
          type: integer
          example: 2
        title:
          type: string
          example: "Dune"
        author:
          type: string
          example: "Frank Herbert"
        status:
          $ref: "#/components/schemas/BookStatus"
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        rating:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        categoryId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, ownerUserId, title, author, status, createdAt, updatedAt]
    CreateBookRequest:
      type: object
      properties:
        title:
          type: string
          example: "Dune"
        author:
          type: string
          example: "Frank Herbert"
        status:
          $ref: "#/components/schemas/BookStatus"
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        rating:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        categoryId:
          type: integer
          nullable: true
      required: [title, author]
    UpdateBookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        status:
          $ref: "#/components/schemas/BookStatus"
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        rating:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        categoryId:
          type: integer
          nullable: true

    # ---------- Goals ----------
    Goal:
      type: object
      properties:
        id:
          type: integer
          example: 3
        childId:
          type: integer
          example: 2
        createdByParentId:
          type: integer
          example: 1
        name:
          type: string
          example: "PlayStation"
        description:
          type: string
          nullable: true
        targetPoints:
          type: integer
          minimum: 1
          example: 1000
        status:
          $ref: "#/components/schemas/GoalStatus"
        createdAt:
          type: string
          format: date-time
      required: [id, childId, createdByParentId, name, targetPoints, status, createdAt]
    GoalWithProgress:
      allOf:
        - $ref: "#/components/schemas/Goal"
        - type: object
          properties:
            pointsAccumulated:
              type: integer
              example: 450
            progressPct:
              type: number
              format: float
              example: 45.0
    CreateGoalRequest:
      type: object
      properties:
        name:
          type: string
          example: "PlayStation"
        description:
          type: string
        targetPoints:
          type: integer
          minimum: 1
          example: 1000
      required: [name, targetPoints]
    UpdateGoalRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        targetPoints:
          type: integer
          minimum: 1

    # ---------- Rewards ----------
...